+++ src/main/java/software/xdev/pmd/model/config/bundled/BundledConfigurationLocation.java
@@ -8,9 +8,9 @@ import org.jetbrains.annotations.Nullable;
 import com.intellij.openapi.project.Project;
 
 import net.sourceforge.pmd.lang.rule.RuleSet;
-import net.sourceforge.pmd.lang.rule.RuleSetLoader;
 import software.xdev.pmd.model.config.ConfigurationLocation;
 import software.xdev.pmd.model.config.ConfigurationType;
+import software.xdev.pmd.model.config.file.pmd.DefaultRuleSetLoaderCreator;
 
 
 public class BundledConfigurationLocation extends ConfigurationLocation
@@ -58,20 +58,21 @@ public class BundledConfigurationLocation extends ConfigurationLocation
 	}
 	
 	@Override
-	public void validate()
+	public void validate(final ClassLoader classLoader)
 	{
 		// always valid
 	}
 	
 	@Nullable
 	@Override
-	protected synchronized RuleSet loadRuleSet()
+	protected synchronized RuleSet loadRuleSet(final ClassLoader ignored)
 	{
-		return new RuleSetLoader().loadFromResource(this.getBundledConfig().getLocation());
+		return DefaultRuleSetLoaderCreator.createAndLoad(rsl -> rsl
+			.loadFromResource(this.getBundledConfig().getLocation()));
 	}
 	
 	@Override
-	protected boolean shouldReloadRuleSet()
+	protected boolean shouldReloadRuleSet(final ClassLoader ignored)
 	{
 		return false;
 	}