+++ src/test/java/software/xdev/pmd/model/config/file/pmd/DefaultRuleSetLoaderCreatorTest.java
@@ -1,0 +1,17 @@
+package software.xdev.pmd.model.config.file.pmd;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+
+class DefaultRuleSetLoaderCreatorTest
+{
+	@Test
+	void checkReflection()
+	{
+		Assertions.assertDoesNotThrow(DefaultRuleSetLoaderCreator::initReflection);
+		
+		Assertions.assertDoesNotThrow(() -> DefaultRuleSetLoaderCreator.createAndLoad(rsl -> rsl
+			.loadFromResource("category/java/security.xml")));
+	}
+}